requirements.txt
# Dependências recomendadas para o projeto LangChain
# Instale com: pip install -r requirements.txt

tiktoken
openai
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
